Harden SqlServer EF mappings for list and unsigned types

+++ FishMMO-Database/FishMMO-DB/Npgsql/EntityConfigurations/Scene/Character/CharacterEntityConfiguration.cs
@@ -27,7 +27,7 @@ namespace FishMMO.Database.SqlServer.Entities
 				.IsRequired()
 				.HasConversion<short>()
 				.HasColumnType("smallint")
-				.HasDefaultValue((short)CharacterSessionState.Offline);
+				.HasDefaultValue(CharacterSessionState.Offline);
 
 			builder.Property(e => e.SessionOwnerServerId)
 				.IsRequired()
